// src/types/payload-types.ts
export interface Media {
  id: string
  alt: string
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
}
// ...
export interface Experience {
  id: string
  name: string
  /**
   * Short description shown on the primary experience list.
   */
  details: {
    root: {
      type: string
      children: {
        type: any
        version: number
        [k: string]: unknown
      }[]
      direction: ('ltr' | 'rtl') | null
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | ''
      indent: number
      version: number
    }
    [k: string]: unknown
  }
  yearRange: string
  position: string
  ctaHref: string
  /**
   * Image of the Company
   */
  companyImage: string | Media
  /**
   * Optional set of related projects (AskMyGov, Directory, etc.).
   */
  projects?:
    | {
        title: string
        href: string
        /**
         * Matches a frontend icon component name, e.g., AskMyGovIcon.
         */
        iconKey: string
        id?: string | null
      }[]
    | null
  updatedAt: string
  createdAt: string
}